Разделил реализацию SqliteFolderRepository на два класса
